+++ TypeContractor/Output/OutputEnumMember.cs
@@ -1,6 +1,6 @@
 namespace TypeContractor.Output;
 
-public record OutputEnumMember(string SourceName, string DestinationName, object DestinationValue)
+public record OutputEnumMember(string SourceName, string DestinationName, object DestinationValue, ObsoleteInfo? Obsolete)
 {
     public override string ToString()
     {